Add config for the server.
- The config will be on when the server is running but in the test we don't need the auto join feature.
- Move the main functions to lib.
